refactor(protocols): split dns.go (1500 → 243 lines) by concern
DNS handler was 1500 lines mixing four separable concerns. Same-package
split — no behaviour or signature changes.

  dns.go (slim)      constants, types, lifecycle, record management
                     (Add/AddPTR/SetDomain/LoadDeviceRecords/LoadDeviceDNSConfig)
  dns_handler.go     HandleQuery / HandleQueryV6 + transport (build/send response,
                     parseDNSLayer, log + validate helpers, SendDNSResponse[V6])
  dns_resolver.go    question resolution (resolveARecord / resolveAAAARecord /
                     resolvePTRRecord), lookupHost, lookupPTR, device selection,
                     PTR parsing, isValidDNSName
  dns_nbstat.go      all NetBIOS NBSTAT-over-DNS code (handlers, response
                     builders, name encoders, service decoders)

go test ./internal/protocols/... passes; golangci-lint clean.
